requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpx==0.28.1
idna==3.10
jinja2==3.1.6
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
